find bug antia

+++ bonus/include/event/event_light_bonus.h
@@ -3,7 +3,8 @@
 
 # include "se_mini_struct_bonus.h"
 
-void	new_ray(t_cam *cam, t_ray *ray, int x, int y);
+// void	new_ray(t_cam *cam, t_ray *ray, int x, int y);
+void		new_ray(t_cam *cam, t_ray *ray, double x, double y);
 double	is_intersect_sphere(t_ray *ray, void *input_sphere, t_ray_vector *i);
 int		is_behind_cam(double t);
